Use structured data in applyFilter
Address one of my FIXMEs by referring to self.shortcuts rather than
analysing data on the html page, which was a silly idea to begin with.

Remove the keywordFilter as this seemed to be its only purpose and all
it did was store id + url + keywords from self.shortcuts

Next we'll need to make sure the search and keyword filters are applied
additively rather than one replacing the other.

We'll also want to do the opposite, and filter the current tag list down
to ones appearing in the current search.